backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
Pillow==10.0.0
opencv-python==4.8.0.74
numpy==1.24.3
requests==2.31.0
openai==1.3.0
anthropic==0.7.8
langchain==0.0.308
langchain-openai==0.0.5
langgraph==0.0.3
pydantic==2.4.2
scikit-image==0.21.0
scikit-learn==1.3.1
torch==2.0.1
torchvision==0.15.2
transformers==4.33.0
redis==5.0.0
SQLAlchemy==2.0.21
gunicorn==21.2.0
aiohttp==3.8.6
httpx==0.24.1
